+++ para/compiler/parser/python/__init__.py
@@ -1,0 +1,6 @@
+# coding=utf-8
+""" Python generated code for the Core Compiler Parser using Antlr4 """
+
+from .ParaLexer import *
+from .ParaListener import *
+from .ParaParser import *