+++ para/preprocessor/__init__.py
@@ -1,0 +1,27 @@
+# coding=utf-8
+"""
+Pre-Processor Module, containing the entire preprocessor including the
+generated code for the preprocessor based on the ParaPreProcessor.g4 grammar
+file
+"""
+import logging
+
+from . import __main__
+from . import ctx
+from . import listener
+from . import logic_tokens
+from . import abc
+from .__main__ import *
+from .logic_tokens import *
+
+__all__ = [
+    'logic_tokens',
+    'listener',
+    'ctx',
+    'abc',
+    *logic_tokens.__all__,
+    *ctx.__all__,
+    *__main__.__all__
+]
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())