+++ para/preprocessor/parser/ParaPreProcessorLexer.py
@@ -1,0 +1,215 @@
+# Generated from ./ParaPreProcessor.g4 by ANTLR 4.9.2
+from antlr4 import *
+from io import StringIO
+import sys
+if sys.version_info[1] > 5:
+    from typing import TextIO
+else:
+    from typing.io import TextIO
+
+
+
+def serializedATN():
+    with StringIO() as buf:
+        buf.write("\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\2\30")
+        buf.write("\u013b\b\1\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7")
+        buf.write("\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r")
+        buf.write("\4\16\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22\t\22\4\23")
+        buf.write("\t\23\4\24\t\24\4\25\t\25\4\26\t\26\4\27\t\27\4\30\t\30")
+        buf.write("\4\31\t\31\4\32\t\32\4\33\t\33\4\34\t\34\4\35\t\35\4\36")
+        buf.write("\t\36\4\37\t\37\4 \t \4!\t!\4\"\t\"\4#\t#\4$\t$\4%\t%")
+        buf.write("\3\2\3\2\7\2N\n\2\f\2\16\2Q\13\2\3\3\3\3\3\3\3\3\3\3\3")
+        buf.write("\3\3\3\3\3\3\3\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\5\3\5")
+        buf.write("\3\5\3\5\3\5\3\5\3\5\3\6\3\6\3\6\3\6\3\7\3\7\3\7\3\7\3")
+        buf.write("\7\3\7\3\b\3\b\3\b\3\b\3\b\3\b\3\b\3\t\3\t\3\t\3\t\3\t")
+        buf.write("\3\t\3\t\3\t\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3")
+        buf.write("\13\3\13\3\13\3\13\3\13\3\13\3\13\3\13\3\13\3\f\3\f\3")
+        buf.write("\f\3\f\3\f\3\f\3\r\3\r\3\r\3\r\3\r\3\r\3\r\3\16\3\16\3")
+        buf.write("\16\3\16\3\16\3\16\3\16\3\17\3\17\3\17\3\17\3\17\3\17")
+        buf.write("\3\17\3\17\3\20\3\20\3\20\3\20\3\20\3\20\3\21\3\21\3\21")
+        buf.write("\7\21\u00bc\n\21\f\21\16\21\u00bf\13\21\3\22\3\22\5\22")
+        buf.write("\u00c3\n\22\3\23\3\23\5\23\u00c7\n\23\3\23\3\23\3\24\3")
+        buf.write("\24\5\24\u00cd\n\24\3\24\3\24\3\25\6\25\u00d2\n\25\r\25")
+        buf.write("\16\25\u00d3\3\26\6\26\u00d7\n\26\r\26\16\26\u00d8\3\27")
+        buf.write("\3\27\3\27\3\27\3\27\3\27\3\27\5\27\u00e2\n\27\3\30\3")
+        buf.write("\30\3\30\3\30\5\30\u00e8\n\30\3\31\3\31\3\31\3\32\3\32")
+        buf.write("\3\32\5\32\u00f0\n\32\3\32\5\32\u00f3\n\32\3\33\3\33\3")
+        buf.write("\33\3\33\6\33\u00f9\n\33\r\33\16\33\u00fa\3\34\3\34\3")
+        buf.write("\35\3\35\3\36\3\36\3\36\3\36\3\36\3\36\3\36\3\36\3\36")
+        buf.write("\3\36\5\36\u010b\n\36\3\37\3\37\3\37\3\37\3\37\3 \3 \3")
+        buf.write("!\3!\3\"\3\"\3\"\3\"\3\"\7\"\u011b\n\"\f\"\16\"\u011e")
+        buf.write("\13\"\3\"\3\"\7\"\u0122\n\"\f\"\16\"\u0125\13\"\3\"\3")
+        buf.write("\"\3\"\3\"\3#\6#\u012c\n#\r#\16#\u012d\3$\3$\5$\u0132")
+        buf.write("\n$\3$\5$\u0135\n$\3%\6%\u0138\n%\r%\16%\u0139\2\2&\3")
+        buf.write("\2\5\3\7\4\t\5\13\6\r\7\17\b\21\t\23\n\25\13\27\f\31\r")
+        buf.write("\33\16\35\17\37\20!\21#\2%\22\'\23)\24+\2-\2/\2\61\2\63")
+        buf.write("\2\65\2\67\29\2;\2=\2?\2A\2C\25E\26G\27I\30\3\2\f\6\2")
+        buf.write("\f\f\17\17$$^^\f\2$$))AA^^cdhhppttvvxx\5\2C\\aac|\3\2")
+        buf.write("\62;\3\2\629\5\2\62;CHch\3\2}}\3\2\177\177\4\2\13\13\"")
+        buf.write("\"\6\2\f\f\17\17\"\"%%\2\u0144\2\5\3\2\2\2\2\7\3\2\2\2")
+        buf.write("\2\t\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2\2\17\3\2\2\2\2\21")
+        buf.write("\3\2\2\2\2\23\3\2\2\2\2\25\3\2\2\2\2\27\3\2\2\2\2\31\3")
+        buf.write("\2\2\2\2\33\3\2\2\2\2\35\3\2\2\2\2\37\3\2\2\2\2!\3\2\2")
+        buf.write("\2\2%\3\2\2\2\2\'\3\2\2\2\2)\3\2\2\2\2C\3\2\2\2\2E\3\2")
+        buf.write("\2\2\2G\3\2\2\2\2I\3\2\2\2\3K\3\2\2\2\5R\3\2\2\2\7[\3")
+        buf.write("\2\2\2\tc\3\2\2\2\13j\3\2\2\2\rn\3\2\2\2\17t\3\2\2\2\21")
+        buf.write("{\3\2\2\2\23\u0083\3\2\2\2\25\u008d\3\2\2\2\27\u0096\3")
+        buf.write("\2\2\2\31\u009c\3\2\2\2\33\u00a3\3\2\2\2\35\u00aa\3\2")
+        buf.write("\2\2\37\u00b2\3\2\2\2!\u00b8\3\2\2\2#\u00c2\3\2\2\2%\u00c4")
+        buf.write("\3\2\2\2\'\u00ca\3\2\2\2)\u00d1\3\2\2\2+\u00d6\3\2\2\2")
+        buf.write("-\u00e1\3\2\2\2/\u00e7\3\2\2\2\61\u00e9\3\2\2\2\63\u00ec")
+        buf.write("\3\2\2\2\65\u00f4\3\2\2\2\67\u00fc\3\2\2\29\u00fe\3\2")
+        buf.write("\2\2;\u010a\3\2\2\2=\u010c\3\2\2\2?\u0111\3\2\2\2A\u0113")
+        buf.write("\3\2\2\2C\u0115\3\2\2\2E\u012b\3\2\2\2G\u0134\3\2\2\2")
+        buf.write("I\u0137\3\2\2\2KO\7%\2\2LN\5E#\2ML\3\2\2\2NQ\3\2\2\2O")
+        buf.write("M\3\2\2\2OP\3\2\2\2P\4\3\2\2\2QO\3\2\2\2RS\5\3\2\2ST\7")
+        buf.write("k\2\2TU\7p\2\2UV\7e\2\2VW\7n\2\2WX\7w\2\2XY\7f\2\2YZ\7")
+        buf.write("g\2\2Z\6\3\2\2\2[\\\5\3\2\2\\]\7f\2\2]^\7g\2\2^_\7h\2")
+        buf.write("\2_`\7k\2\2`a\7p\2\2ab\7g\2\2b\b\3\2\2\2cd\5\3\2\2de\7")
+        buf.write("w\2\2ef\7p\2\2fg\7f\2\2gh\7g\2\2hi\7h\2\2i\n\3\2\2\2j")
+        buf.write("k\5\3\2\2kl\7k\2\2lm\7h\2\2m\f\3\2\2\2no\5\3\2\2op\7g")
+        buf.write("\2\2pq\7n\2\2qr\7u\2\2rs\7g\2\2s\16\3\2\2\2tu\5\3\2\2")
+        buf.write("uv\7k\2\2vw\7h\2\2wx\7f\2\2xy\7g\2\2yz\7h\2\2z\20\3\2")
+        buf.write("\2\2{|\5\3\2\2|}\7k\2\2}~\7h\2\2~\177\7p\2\2\177\u0080")
+        buf.write("\7f\2\2\u0080\u0081\7g\2\2\u0081\u0082\7h\2\2\u0082\22")
+        buf.write("\3\2\2\2\u0083\u0084\5\3\2\2\u0084\u0085\7g\2\2\u0085")
+        buf.write("\u0086\7n\2\2\u0086\u0087\7k\2\2\u0087\u0088\7h\2\2\u0088")
+        buf.write("\u0089\7p\2\2\u0089\u008a\7f\2\2\u008a\u008b\7g\2\2\u008b")
+        buf.write("\u008c\7h\2\2\u008c\24\3\2\2\2\u008d\u008e\5\3\2\2\u008e")
+        buf.write("\u008f\7g\2\2\u008f\u0090\7n\2\2\u0090\u0091\7k\2\2\u0091")
+        buf.write("\u0092\7h\2\2\u0092\u0093\7f\2\2\u0093\u0094\7g\2\2\u0094")
+        buf.write("\u0095\7h\2\2\u0095\26\3\2\2\2\u0096\u0097\5\3\2\2\u0097")
+        buf.write("\u0098\7g\2\2\u0098\u0099\7n\2\2\u0099\u009a\7k\2\2\u009a")
+        buf.write("\u009b\7h\2\2\u009b\30\3\2\2\2\u009c\u009d\5\3\2\2\u009d")
+        buf.write("\u009e\7g\2\2\u009e\u009f\7p\2\2\u009f\u00a0\7f\2\2\u00a0")
+        buf.write("\u00a1\7k\2\2\u00a1\u00a2\7h\2\2\u00a2\32\3\2\2\2\u00a3")
+        buf.write("\u00a4\5\3\2\2\u00a4\u00a5\7g\2\2\u00a5\u00a6\7t\2\2\u00a6")
+        buf.write("\u00a7\7t\2\2\u00a7\u00a8\7q\2\2\u00a8\u00a9\7t\2\2\u00a9")
+        buf.write("\34\3\2\2\2\u00aa\u00ab\5\3\2\2\u00ab\u00ac\7r\2\2\u00ac")
+        buf.write("\u00ad\7t\2\2\u00ad\u00ae\7c\2\2\u00ae\u00af\7i\2\2\u00af")
+        buf.write("\u00b0\7o\2\2\u00b0\u00b1\7c\2\2\u00b1\36\3\2\2\2\u00b2")
+        buf.write("\u00b3\5\3\2\2\u00b3\u00b4\7n\2\2\u00b4\u00b5\7k\2\2\u00b5")
+        buf.write("\u00b6\7p\2\2\u00b6\u00b7\7g\2\2\u00b7 \3\2\2\2\u00b8")
+        buf.write("\u00bd\5#\22\2\u00b9\u00bc\5#\22\2\u00ba\u00bc\59\35\2")
+        buf.write("\u00bb\u00b9\3\2\2\2\u00bb\u00ba\3\2\2\2\u00bc\u00bf\3")
+        buf.write("\2\2\2\u00bd\u00bb\3\2\2\2\u00bd\u00be\3\2\2\2\u00be\"")
+        buf.write("\3\2\2\2\u00bf\u00bd\3\2\2\2\u00c0\u00c3\5\67\34\2\u00c1")
+        buf.write("\u00c3\5;\36\2\u00c2\u00c0\3\2\2\2\u00c2\u00c1\3\2\2\2")
+        buf.write("\u00c3$\3\2\2\2\u00c4\u00c6\7>\2\2\u00c5\u00c7\5+\26\2")
+        buf.write("\u00c6\u00c5\3\2\2\2\u00c6\u00c7\3\2\2\2\u00c7\u00c8\3")
+        buf.write("\2\2\2\u00c8\u00c9\7@\2\2\u00c9&\3\2\2\2\u00ca\u00cc\7")
+        buf.write("$\2\2\u00cb\u00cd\5+\26\2\u00cc\u00cb\3\2\2\2\u00cc\u00cd")
+        buf.write("\3\2\2\2\u00cd\u00ce\3\2\2\2\u00ce\u00cf\7$\2\2\u00cf")
+        buf.write("(\3\2\2\2\u00d0\u00d2\59\35\2\u00d1\u00d0\3\2\2\2\u00d2")
+        buf.write("\u00d3\3\2\2\2\u00d3\u00d1\3\2\2\2\u00d3\u00d4\3\2\2\2")
+        buf.write("\u00d4*\3\2\2\2\u00d5\u00d7\5-\27\2\u00d6\u00d5\3\2\2")
+        buf.write("\2\u00d7\u00d8\3\2\2\2\u00d8\u00d6\3\2\2\2\u00d8\u00d9")
+        buf.write("\3\2\2\2\u00d9,\3\2\2\2\u00da\u00e2\n\2\2\2\u00db\u00e2")
+        buf.write("\5/\30\2\u00dc\u00dd\7^\2\2\u00dd\u00e2\7\f\2\2\u00de")
+        buf.write("\u00df\7^\2\2\u00df\u00e0\7\17\2\2\u00e0\u00e2\7\f\2\2")
+        buf.write("\u00e1\u00da\3\2\2\2\u00e1\u00db\3\2\2\2\u00e1\u00dc\3")
+        buf.write("\2\2\2\u00e1\u00de\3\2\2\2\u00e2.\3\2\2\2\u00e3\u00e8")
+        buf.write("\5\61\31\2\u00e4\u00e8\5\63\32\2\u00e5\u00e8\5\65\33\2")
+        buf.write("\u00e6\u00e8\5;\36\2\u00e7\u00e3\3\2\2\2\u00e7\u00e4\3")
+        buf.write("\2\2\2\u00e7\u00e5\3\2\2\2\u00e7\u00e6\3\2\2\2\u00e8\60")
+        buf.write("\3\2\2\2\u00e9\u00ea\7^\2\2\u00ea\u00eb\t\3\2\2\u00eb")
+        buf.write("\62\3\2\2\2\u00ec\u00ed\7^\2\2\u00ed\u00ef\5? \2\u00ee")
+        buf.write("\u00f0\5? \2\u00ef\u00ee\3\2\2\2\u00ef\u00f0\3\2\2\2\u00f0")
+        buf.write("\u00f2\3\2\2\2\u00f1\u00f3\5? \2\u00f2\u00f1\3\2\2\2\u00f2")
+        buf.write("\u00f3\3\2\2\2\u00f3\64\3\2\2\2\u00f4\u00f5\7^\2\2\u00f5")
+        buf.write("\u00f6\7z\2\2\u00f6\u00f8\3\2\2\2\u00f7\u00f9\5A!\2\u00f8")
+        buf.write("\u00f7\3\2\2\2\u00f9\u00fa\3\2\2\2\u00fa\u00f8\3\2\2\2")
+        buf.write("\u00fa\u00fb\3\2\2\2\u00fb\66\3\2\2\2\u00fc\u00fd\t\4")
+        buf.write("\2\2\u00fd8\3\2\2\2\u00fe\u00ff\t\5\2\2\u00ff:\3\2\2\2")
+        buf.write("\u0100\u0101\7^\2\2\u0101\u0102\7w\2\2\u0102\u0103\3\2")
+        buf.write("\2\2\u0103\u010b\5=\37\2\u0104\u0105\7^\2\2\u0105\u0106")
+        buf.write("\7W\2\2\u0106\u0107\3\2\2\2\u0107\u0108\5=\37\2\u0108")
+        buf.write("\u0109\5=\37\2\u0109\u010b\3\2\2\2\u010a\u0100\3\2\2\2")
+        buf.write("\u010a\u0104\3\2\2\2\u010b<\3\2\2\2\u010c\u010d\5A!\2")
+        buf.write("\u010d\u010e\5A!\2\u010e\u010f\5A!\2\u010f\u0110\5A!\2")
+        buf.write("\u0110>\3\2\2\2\u0111\u0112\t\6\2\2\u0112@\3\2\2\2\u0113")
+        buf.write("\u0114\t\7\2\2\u0114B\3\2\2\2\u0115\u0116\7c\2\2\u0116")
+        buf.write("\u0117\7u\2\2\u0117\u0118\7o\2\2\u0118\u011c\3\2\2\2\u0119")
+        buf.write("\u011b\n\b\2\2\u011a\u0119\3\2\2\2\u011b\u011e\3\2\2\2")
+        buf.write("\u011c\u011a\3\2\2\2\u011c\u011d\3\2\2\2\u011d\u011f\3")
+        buf.write("\2\2\2\u011e\u011c\3\2\2\2\u011f\u0123\7}\2\2\u0120\u0122")
+        buf.write("\n\t\2\2\u0121\u0120\3\2\2\2\u0122\u0125\3\2\2\2\u0123")
+        buf.write("\u0121\3\2\2\2\u0123\u0124\3\2\2\2\u0124\u0126\3\2\2\2")
+        buf.write("\u0125\u0123\3\2\2\2\u0126\u0127\7\177\2\2\u0127\u0128")
+        buf.write("\3\2\2\2\u0128\u0129\b\"\2\2\u0129D\3\2\2\2\u012a\u012c")
+        buf.write("\t\n\2\2\u012b\u012a\3\2\2\2\u012c\u012d\3\2\2\2\u012d")
+        buf.write("\u012b\3\2\2\2\u012d\u012e\3\2\2\2\u012eF\3\2\2\2\u012f")
+        buf.write("\u0131\7\17\2\2\u0130\u0132\7\f\2\2\u0131\u0130\3\2\2")
+        buf.write("\2\u0131\u0132\3\2\2\2\u0132\u0135\3\2\2\2\u0133\u0135")
+        buf.write("\7\f\2\2\u0134\u012f\3\2\2\2\u0134\u0133\3\2\2\2\u0135")
+        buf.write("H\3\2\2\2\u0136\u0138\n\13\2\2\u0137\u0136\3\2\2\2\u0138")
+        buf.write("\u0139\3\2\2\2\u0139\u0137\3\2\2\2\u0139\u013a\3\2\2\2")
+        buf.write("\u013aJ\3\2\2\2\27\2O\u00bb\u00bd\u00c2\u00c6\u00cc\u00d3")
+        buf.write("\u00d8\u00e1\u00e7\u00ef\u00f2\u00fa\u010a\u011c\u0123")
+        buf.write("\u012d\u0131\u0134\u0139\3\b\2\2")
+        return buf.getvalue()
+
+
+class ParaPreProcessorLexer(Lexer):
+
+    atn = ATNDeserializer().deserialize(serializedATN())
+
+    decisionsToDFA = [ DFA(ds, i) for i, ds in enumerate(atn.decisionToState) ]
+
+    Include = 1
+    Define = 2
+    Undefine = 3
+    If = 4
+    Else = 5
+    IfDefined = 6
+    IfNotDefined = 7
+    ElIfNotDefined = 8
+    ElIfDefined = 9
+    ElseIf = 10
+    EndIf = 11
+    Error = 12
+    Pragma = 13
+    Line = 14
+    Identifier = 15
+    LibStringLiteral = 16
+    StringLiteral = 17
+    DigitSequence = 18
+    AsmBlock = 19
+    Whitespace = 20
+    Newline = 21
+    NonPreProcessorItemSequence = 22
+
+    channelNames = [ u"DEFAULT_TOKEN_CHANNEL", u"HIDDEN" ]
+
+    modeNames = [ "DEFAULT_MODE" ]
+
+    literalNames = [ "<INVALID>",
+ ]
+
+    symbolicNames = [ "<INVALID>",
+            "Include", "Define", "Undefine", "If", "Else", "IfDefined", 
+            "IfNotDefined", "ElIfNotDefined", "ElIfDefined", "ElseIf", "EndIf", 
+            "Error", "Pragma", "Line", "Identifier", "LibStringLiteral", 
+            "StringLiteral", "DigitSequence", "AsmBlock", "Whitespace", 
+            "Newline", "NonPreProcessorItemSequence" ]
+
+    ruleNames = [ "PreProcessorBegin", "Include", "Define", "Undefine", 
+                  "If", "Else", "IfDefined", "IfNotDefined", "ElIfNotDefined", 
+                  "ElIfDefined", "ElseIf", "EndIf", "Error", "Pragma", "Line", 
+                  "Identifier", "IdentifierNondigit", "LibStringLiteral", 
+                  "StringLiteral", "DigitSequence", "SCharSequence", "SChar", 
+                  "EscapeSequence", "SimpleEscapeSequence", "OctalEscapeSequence", 
+                  "HexadecimalEscapeSequence", "Nondigit", "Digit", "UniversalCharacterName", 
+                  "HexQuad", "OctalDigit", "HexadecimalDigit", "AsmBlock", 
+                  "Whitespace", "Newline", "NonPreProcessorItemSequence" ]
+
+    grammarFileName = "ParaPreProcessor.g4"
+
+    def __init__(self, input=None, output:TextIO = sys.stdout):
+        super().__init__(input, output)
+        self.checkVersion("4.9.2")
+        self._interp = LexerATNSimulator(self, self.atn, self.decisionsToDFA, PredictionContextCache())
+        self._actions = None
+        self._predicates = None
+
+