+++ para/util/c_compiler.py
@@ -1,0 +1,2 @@
+# coding=utf-8
+""" C-Compiler related handling functions """