+++ pytest/tests/test_utils.py
@@ -1,12 +1,11 @@
 # coding=utf-8
 """
-Test for the utility functions in Para-C (test_util.py and decorators.py)
+Test for the utility functions in Para (test_util.py and decorators.py)
 """
 from pathlib import Path
-
 import pytest
 
-from parac import util
+from para import util
 
 
 class TestValidateFileEnding:
@@ -61,7 +60,7 @@ class TestGetRelativeFileName:
         else:
             assert False
 
-    def test_simple_entry_file_path(self):
+    def test_current_directory_file(self):
         name = util.get_relative_file_name(
             file_name="name.para",
             file_path="/usr/name/files/name.para",